+++ gameserver/src/main/java/com/lhf/game/map/RestArea.java
@@ -20,6 +20,13 @@ import com.lhf.game.enums.Attributes;
 import com.lhf.game.enums.EquipmentSlots;
 import com.lhf.game.lewd.LewdProduct;
 import com.lhf.game.lewd.VrijPartij;
+import com.lhf.game.map.SubArea.RoundThread;
+import com.lhf.game.map.SubArea.SubAreaBuilder;
+import com.lhf.game.map.SubArea.SubAreaCommandHandler;
+import com.lhf.game.map.SubArea.SubAreaExitHandler;
+import com.lhf.game.map.SubArea.SubAreaSayHandler;
+import com.lhf.game.map.SubArea.SubAreaSeeHandler;
+import com.lhf.game.map.SubArea.SubAreaSort;
 import com.lhf.messages.Command;
 import com.lhf.messages.CommandChainHandler;
 import com.lhf.messages.CommandContext;
