+++ gameserver/src/main/java/com/lhf/game/map/Room.java
@@ -5,7 +5,6 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumMap;
-import java.util.EnumSet;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
@@ -18,16 +17,11 @@ import java.util.StringJoiner;
 import java.util.TreeSet;
 import java.util.UUID;
 import java.util.function.Consumer;
-import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.regex.PatternSyntaxException;
 
-import com.lhf.Examinable;
 import com.lhf.game.EntityEffect;
-import com.lhf.game.ItemContainer;
-import com.lhf.game.LockableItemContainer;
 import com.lhf.game.creature.ICreature;
 import com.lhf.game.creature.IMonster;
 import com.lhf.game.creature.INonPlayerCharacter;
@@ -36,53 +30,25 @@ import com.lhf.game.creature.Player;
 import com.lhf.game.creature.conversation.ConversationManager;
 import com.lhf.game.creature.intelligence.AIRunner;
 import com.lhf.game.creature.statblock.StatblockManager;
-import com.lhf.game.item.InteractObject;
 import com.lhf.game.item.Item;
-import com.lhf.game.item.Takeable;
-import com.lhf.game.item.Usable;
 import com.lhf.game.item.concrete.Corpse;
 import com.lhf.game.map.Area.AreaBuilder.PostBuildRoomOperations;
 import com.lhf.game.map.SubArea.SubAreaBuilder;
-import com.lhf.game.map.SubArea.SubAreaSort;
-import com.lhf.messages.Command;
 import com.lhf.messages.CommandChainHandler;
 import com.lhf.messages.CommandContext;
-import com.lhf.messages.CommandContext.Reply;
-import com.lhf.messages.events.BadMessageEvent;
-import com.lhf.messages.events.BadMessageEvent.BadMessageType;
-import com.lhf.messages.events.BadSpeakingTargetEvent;
-import com.lhf.messages.events.BadTargetSelectedEvent;
-import com.lhf.messages.events.BadTargetSelectedEvent.BadTargetOption;
 import com.lhf.messages.events.CreatureDiedEvent;
-import com.lhf.messages.events.ItemDroppedEvent;
-import com.lhf.messages.events.ItemDroppedEvent.DropType;
-import com.lhf.messages.events.ItemInteractionEvent;
-import com.lhf.messages.events.ItemInteractionEvent.InteractOutMessageType;
-import com.lhf.messages.events.ItemNotPossessedEvent;
-import com.lhf.messages.events.ItemTakenEvent;
-import com.lhf.messages.events.ItemTakenEvent.TakeOutType;
-import com.lhf.messages.events.ItemUsedEvent;
-import com.lhf.messages.events.ItemUsedEvent.UseOutMessageOption;
 import com.lhf.messages.events.RoomAffectedEvent;
 import com.lhf.messages.events.RoomEnteredEvent;
 import com.lhf.messages.events.RoomExitedEvent;
 import com.lhf.messages.events.SeeEvent;
-import com.lhf.messages.events.SpeakingEvent;
 import com.lhf.messages.in.AMessageType;
-import com.lhf.messages.in.DropMessage;
-import com.lhf.messages.in.InteractMessage;
-import com.lhf.messages.in.SayMessage;
-import com.lhf.messages.in.SeeMessage;
-import com.lhf.messages.in.TakeMessage;
-import com.lhf.messages.in.UseMessage;
-import com.lhf.server.client.CommandInvoker;
 import com.lhf.server.client.user.UserID;
 
 public class Room implements Area {
     private final GameEventProcessorID gameEventProcessorID = new GameEventProcessorID();
     private final UUID uuid = gameEventProcessorID.getUuid();
     protected final Logger logger;
-    private final List<Item> items;
+    final List<Item> items;
     private final String description;
     private final String name;
     private final NavigableSet<SubArea> subAreas;
@@ -305,16 +271,7 @@ public class Room implements Area {
     }
 
     protected Map<AMessageType, CommandHandler> buildCommands() {
-        Map<AMessageType, CommandHandler> cmds = new EnumMap<>(AMessageType.class);
-        cmds.put(AMessageType.SAY, new SayHandler());
-        cmds.put(AMessageType.SEE, new SeeHandler());
-        cmds.put(AMessageType.DROP, new DropHandler());
-        cmds.put(AMessageType.USE, new UseHandler());
-        cmds.put(AMessageType.CAST, new CastHandler());
-        cmds.put(AMessageType.ATTACK, new AttackHandler());
-        cmds.put(AMessageType.REST, new RestHandler());
-        cmds.put(AMessageType.INTERACT, new InteractHandler());
-        cmds.put(AMessageType.TAKE, new TakeHandler());
+        Map<AMessageType, CommandHandler> cmds = new EnumMap<>(Area.AreaCommandHandler.areaCommandHandlers);
         return cmds;
     }
 
@@ -612,667 +569,6 @@ public class Room implements Area {
             ctx.setArea(this);
         }
         return ctx;
-    }
-
-    public interface RoomCommandHandler extends ICreature.CreatureCommandHandler {
-        final static Predicate<CommandContext> defaultRoomPredicate = ICreature.CreatureCommandHandler.defaultCreaturePredicate
-                .and(ctx -> ctx.getArea() != null);
-        final static Predicate<CommandContext> defaultNoBattlePredicate = RoomCommandHandler.defaultRoomPredicate
-                .and(ctx -> !ctx.getCreature().isInBattle());
-        final static String inBattleString = "You appear to be in a fight, so you cannot do that.";
-    }
-
-    protected class AttackHandler implements RoomCommandHandler {
-        private final static String helpString = new StringJoiner(" ")
-                .add("\"attack [name]\"").add("Attacks a creature").add("\r\n")
-                .add("\"attack [name] with [weapon]\"").add("Attack the named creature with a weapon that you have.")
-                .add("In the unlikely event that either the creature or the weapon's name contains 'with', enclose the name in quotation marks.")
-                .toString();
-        private final static Predicate<CommandContext> enabledPredicate = AttackHandler.defaultRoomPredicate
-                .and(ctx -> {
-                    Area room = ctx.getArea();
-                    if (!room.hasSubAreaSort(SubAreaSort.BATTLE)) {
-                        room.log(Level.WARNING, () -> String.format("No battle manager for room: %s", room.getName()));
-                        return false;
-                    }
-                    return room.getCreatures().size() > 1;
-                });
-
-        @Override
-        public AMessageType getHandleType() {
-            return AMessageType.ATTACK;
-        }
-
-        @Override
-        public Optional<String> getHelp(CommandContext ctx) {
-            return Optional.of(AttackHandler.helpString);
-        }
-
-        @Override
-        public Predicate<CommandContext> getEnabledPredicate() {
-            return AttackHandler.enabledPredicate;
-        }
-
-        @Override
-        public Reply handleCommand(CommandContext ctx, Command cmd) {
-            if (cmd == null || cmd.getType() != AMessageType.ATTACK) {
-                return ctx.failhandle();
-            }
-            ctx = Room.this.addSelfToContext(ctx);
-            if (ctx.getCreature() == null) {
-                ctx.receive(BadMessageEvent.getBuilder().setBadMessageType(BadMessageType.CREATURES_ONLY)
-                        .setHelps(ctx.getHelps()).setCommand(cmd).Build());
-                return ctx.handled();
-            }
-            final SubArea subArea = Room.this.getSubAreaForSort(SubAreaSort.BATTLE);
-            if (subArea == null) {
-                this.log(Level.WARNING, "No battle sub area found!");
-                return ctx.failhandle();
-            }
-            return subArea.handleChain(ctx, cmd);
-        }
-
-        @Override
-        public CommandChainHandler getChainHandler() {
-            return Room.this;
-        }
-
-    }
-
-    protected class CastHandler implements RoomCommandHandler {
-
-        @Override
-        public AMessageType getHandleType() {
-            return AMessageType.CAST;
-        }
-
-        @Override
-        public Optional<String> getHelp(CommandContext ctx) {
-            return Optional.empty();
-        }
-
-        @Override
-        public Predicate<CommandContext> getEnabledPredicate() {
-            return CastHandler.defaultRoomPredicate;
-        }
-
-        @Override
-        public Reply handleCommand(CommandContext ctx, Command cmd) {
-            if (ctx.getCreature() == null) {
-                ctx.receive(BadMessageEvent.getBuilder().setBadMessageType(BadMessageType.CREATURES_ONLY)
-                        .setHelps(ctx.getHelps()).setCommand(cmd).Build());
-                return ctx.handled();
-            }
-            return ctx.failhandle(); // let a successor (ThirdPower) handle it
-        }
-
-        @Override
-        public CommandChainHandler getChainHandler() {
-            return Room.this;
-        }
-
-    }
-
-    protected class TakeHandler implements RoomCommandHandler {
-        private final static Predicate<CommandContext> enabledPredicate = InteractHandler.defaultNoBattlePredicate
-                .and(ctx -> ctx.getArea()
-                        .filterItems(EnumSet.of(ItemFilters.TYPE), null, null, null, Takeable.class, null).size() > 0);
-        private final static String helpString = new StringJoiner(" ").add("\"take [item]\"")
-                .add("Take an item from the room and add it to your inventory.\n")
-                .add("\"take [item] from \"[someone]'s corpse\"")
-                .add("Take an item from a container of some kind, just double-quote the container name")
-                .toString();
-
-        @Override
-        public AMessageType getHandleType() {
-            return AMessageType.TAKE;
-        }
-
-        @Override
-        public Optional<String> getHelp(CommandContext ctx) {
-            if (ctx == null || ctx.getCreature() == null) {
-                return Optional.empty();
-            }
-            return Optional
-                    .of(ctx.getCreature().isInBattle() ? TakeHandler.inBattleString : TakeHandler.helpString);
-        }
-
-        @Override
-        public Predicate<CommandContext> getEnabledPredicate() {
-            return TakeHandler.enabledPredicate;
-        }
-
-        @Override
-        public Reply handleCommand(CommandContext ctx, Command cmd) {
-            if (cmd != null && cmd.getType() == AMessageType.TAKE && cmd instanceof TakeMessage tMessage) {
-                if (ctx.getCreature() == null) {
-                    ctx.receive(BadMessageEvent.getBuilder().setBadMessageType(BadMessageType.CREATURES_ONLY)
-                            .setHelps(ctx.getHelps()).setCommand(tMessage).Build());
-                    return ctx.handled();
-                }
-
-                ItemTakenEvent.Builder takeOutMessage = ItemTakenEvent.getBuilder();
-
-                ItemContainer container = Room.this;
-                takeOutMessage.setSource(Room.this);
-                Optional<String> containerName = tMessage.fromContainer();
-                if (containerName.isPresent()) {
-                    takeOutMessage.setSource(containerName.orElse(null));
-                    Optional<ItemContainer> foundContainer = Room.this.items.stream()
-                            .filter(item -> item != null && item instanceof ItemContainer
-                                    && item.checkName(containerName.get().replaceAll("^\"|\"$", "")))
-                            .map(item -> (ItemContainer) item).findAny();
-                    if (foundContainer.isEmpty()) {
-                        ctx.receive(takeOutMessage.setSubType(TakeOutType.BAD_CONTAINER).Build());
-                        return ctx.handled();
-                    }
-                    if (foundContainer.get() instanceof LockableItemContainer liCon) {
-                        if (!liCon.canAccess(ctx.getCreature())) {
-                            ctx.receive(takeOutMessage.setSubType(TakeOutType.LOCKED_CONTAINER).Build());
-                            return ctx.handled();
-                        }
-                        container = liCon.getBypass();
-                    } else {
-                        container = foundContainer.get();
-                    }
-                }
-
-                for (String thing : tMessage.getDirects()) {
-                    takeOutMessage.setAttemptedName(thing);
-                    if (thing.length() < 3) {
-                        ctx.receive(takeOutMessage.setSubType(TakeOutType.SHORT).Build());
-                        continue;
-                    }
-                    if (thing.matches("[^ a-zA-Z_-]+") || thing.contains("*")) {
-                        ctx.receive(takeOutMessage.setSubType(TakeOutType.INVALID).Build());
-                        continue;
-                    }
-                    try {
-                        Optional<Item> maybeItem = container.getItems().stream()
-                                .filter(item -> item.CheckNameRegex(thing, 3))
-                                .findAny();
-                        if (maybeItem.isEmpty()) {
-                            if (thing.equalsIgnoreCase("all") || thing.equalsIgnoreCase("everything")) {
-                                ctx.receive(takeOutMessage.setSubType(TakeOutType.GREEDY).Build());
-                            } else {
-                                ctx.receive(takeOutMessage.setSubType(TakeOutType.NOT_FOUND).Build());
-                            }
-                            continue;
-                        }
-                        Item item = maybeItem.get();
-                        takeOutMessage.setItem(item);
-                        if (item instanceof Takeable takeableItem) {
-                            ctx.getCreature().addItem(takeableItem);
-                            container.removeItem(takeableItem);
-                            ctx.receive(takeOutMessage.setSubType(TakeOutType.FOUND_TAKEN).Build());
-                            continue;
-                        }
-                        ctx.receive(takeOutMessage.setSubType(TakeOutType.NOT_TAKEABLE).Build());
-                    } catch (PatternSyntaxException pse) {
-                        pse.printStackTrace();
-                        ctx.receive(takeOutMessage.setSubType(TakeOutType.UNCLEVER).Build());
-                    }
-                }
-                while (container instanceof LockableItemContainer.Bypass bypass) {
-                    container = bypass.getOrigin();
-                }
-                if (container instanceof LockableItemContainer liCon && liCon instanceof Item liConItem) {
-                    if (liCon.isRemoveOnEmpty() && liCon.isEmpty()) {
-                        Room.this.removeItem(liConItem);
-                    }
-                }
-                return ctx.handled();
-            }
-            return ctx.failhandle();
-        }
-
-        @Override
-        public CommandChainHandler getChainHandler() {
-            return Room.this;
-        }
-    }
-
-    protected class InteractHandler implements RoomCommandHandler {
-        private final static Predicate<CommandContext> enabledPredicate = InteractHandler.defaultNoBattlePredicate
-                .and(ctx -> ctx.getArea()
-                        .filterItems(EnumSet.of(ItemFilters.TYPE), null, null, null, InteractObject.class, null)
-                        .size() > 0);
-        private final static String helpString = "\"interact [item]\" Certain items in the room may be interactable. Like \"interact lever\"";
-
-        @Override
-        public AMessageType getHandleType() {
-            return AMessageType.INTERACT;
-        }
-
-        @Override
-        public Optional<String> getHelp(CommandContext ctx) {
-            if (ctx == null || ctx.getCreature() == null) {
-                return Optional.empty();
-            }
-            return Optional
-                    .of(ctx.getCreature().isInBattle() ? InteractHandler.inBattleString : InteractHandler.helpString);
-        }
-
-        @Override
-        public Predicate<CommandContext> getEnabledPredicate() {
-            return InteractHandler.enabledPredicate;
-        }
-
-        @Override
-        public Reply handleCommand(CommandContext ctx, Command cmd) {
-            if (cmd != null && cmd.getType() == AMessageType.INTERACT && cmd instanceof InteractMessage intMessage) {
-                if (ctx.getCreature() == null) {
-                    ctx.receive(BadMessageEvent.getBuilder().setBadMessageType(BadMessageType.CREATURES_ONLY)
-                            .setHelps(ctx.getHelps()).setCommand(cmd).Build());
-                    return ctx.handled();
-                }
-                String name = intMessage.getObject();
-                List<Item> matches = Room.this.getItems().stream()
-                        .filter(ro -> ro != null && ro.CheckNameRegex(name, 3)).toList();
-
-                if (matches.size() == 1) {
-                    Item ro = matches.get(0);
-                    if (ro instanceof InteractObject) {
-                        InteractObject ex = (InteractObject) ro;
-                        ctx.receive(ex.doUseAction(ctx.getCreature()));
-                    } else {
-                        ctx.receive(ItemInteractionEvent.getBuilder().setTaggable(ro)
-                                .setSubType(InteractOutMessageType.CANNOT).Build());
-                    }
-                    return ctx.handled();
-                }
-                List<InteractObject> interactables = Room.this.getItems().stream()
-                        .filter(ro -> ro != null && ro.checkVisibility() && ro instanceof InteractObject)
-                        .map(ro -> (InteractObject) ro).toList();
-                ctx.receive(BadTargetSelectedEvent.getBuilder().setBde(BadTargetOption.UNCLEAR).setBadTarget(name)
-                        .setPossibleTargets(interactables).Build());
-                return ctx.handled();
-            }
-            return ctx.failhandle();
-        }
-
-        @Override
-        public CommandChainHandler getChainHandler() {
-            return Room.this;
-        }
-    }
-
-    protected class DropHandler implements RoomCommandHandler {
-        private static final String helpString = "\"drop [itemname]\" Drop an item that you have. Like \"drop longsword\"";
-        private static final Predicate<CommandContext> enabledPredicate = DropHandler.defaultRoomPredicate
-                .and(ctx -> ctx.getCreature().getItems().size() > 1);
-
-        @Override
-        public AMessageType getHandleType() {
-            return AMessageType.DROP;
-        }
-
-        @Override
-        public Optional<String> getHelp(CommandContext ctx) {
-            return Optional.of(DropHandler.helpString);
-        }
-
-        @Override
-        public Predicate<CommandContext> getEnabledPredicate() {
-            return DropHandler.enabledPredicate;
-        }
-
-        @Override
-        public Reply handleCommand(CommandContext ctx, Command cmd) {
-            if (cmd != null && cmd.getType() == AMessageType.DROP && cmd instanceof DropMessage dMessage) {
-                if (ctx.getCreature() == null) {
-                    ctx.receive(BadMessageEvent.getBuilder().setBadMessageType(BadMessageType.CREATURES_ONLY)
-                            .setHelps(ctx.getHelps()).setCommand(dMessage).Build());
-                    return ctx.handled();
-                }
-                ItemDroppedEvent.Builder dOutMessage = ItemDroppedEvent.getBuilder();
-                if (dMessage.getDirects().size() == 0) {
-                    ctx.receive(dOutMessage.setDropType(DropType.NO_ITEM));
-                    return ctx.handled();
-                }
-
-                ItemContainer container = Room.this;
-                dOutMessage.setDestination(Room.this.getName());
-                Optional<String> containerName = dMessage.inContainer();
-                if (containerName.isPresent()) {
-                    // takeOutMessage.setSource(containerName.orElse(null));
-                    dOutMessage.setDestination(containerName.get());
-                    Optional<ItemContainer> foundContainer = Room.this.items.stream()
-                            .filter(item -> item != null && item instanceof ItemContainer
-                                    && item.checkName(containerName.get().replaceAll("^\"|\"$", "")))
-                            .map(item -> (ItemContainer) item).findAny();
-                    if (foundContainer.isEmpty()) {
-                        ctx.receive(dOutMessage.setDropType(DropType.BAD_CONTAINER));
-                        return ctx.handled();
-                    } else if (foundContainer.get() instanceof LockableItemContainer liCon) {
-                        if (!liCon.canAccess(ctx.getCreature())) {
-                            ctx.receive(dOutMessage.setDropType(DropType.LOCKED_CONTAINER));
-                            return ctx.handled();
-                        }
-                        container = liCon.getBypass();
-                    } else {
-                        container = foundContainer.get();
-                    }
-                }
-                dOutMessage.setDestination(container.getName());
-
-                for (String itemName : dMessage.getDirects()) {
-                    Optional<Item> maybeTakeable = ctx.getCreature().removeItem(itemName);
-                    if (maybeTakeable.isEmpty()) {
-                        ctx.receive(ItemNotPossessedEvent.getBuilder().setItemType(Item.class.getSimpleName())
-                                .setItemName(itemName).Build());
-                        continue;
-                    }
-                    Item takeable = maybeTakeable.get();
-                    container.addItem(takeable);
-                    ctx.receive(dOutMessage.setDropType(DropType.SUCCESS).setItem(takeable)
-                            .setDestination(container.getName()).Build());
-                }
-                return ctx.handled();
-            }
-            return ctx.failhandle();
-        }
-
-        @Override
-        public CommandChainHandler getChainHandler() {
-            return Room.this;
-        }
-    }
-
-    // only used to examine items and creatures in this room
-    protected class SeeHandler implements RoomCommandHandler {
-        private final static String helpString = new StringJoiner(" ")
-                .add("\"see\"").add("Will give you some information about your surroundings.\r\n")
-                .add("\"see [name]\"").add("May tell you more about the object with that name.")
-                .toString();
-
-        @Override
-        public AMessageType getHandleType() {
-            return AMessageType.SEE;
-        }
-
-        @Override
-        public Optional<String> getHelp(CommandContext ctx) {
-            return Optional.of(SeeHandler.helpString);
-        }
-
-        @Override
-        public Predicate<CommandContext> getEnabledPredicate() {
-            return SeeHandler.defaultRoomPredicate;
-        }
-
-        @Override
-        public Reply handleCommand(CommandContext ctx, Command cmd) {
-            if (cmd != null && cmd.getType() == AMessageType.SEE && cmd instanceof SeeMessage sMessage) {
-                if (sMessage.getThing() != null && !sMessage.getThing().isBlank()) {
-                    String name = sMessage.getThing();
-                    Collection<ICreature> found = Room.this.getCreaturesLike(name);
-                    // we should be able to see people in a fight
-                    if (found.size() == 1) {
-                        ArrayList<ICreature> foundList = new ArrayList<ICreature>(found);
-                        ctx.receive(((SeeEvent.Builder) foundList.get(0).produceMessage().copyBuilder())
-                                .addExtraInfo("They are in the room with you. ").Build());
-                        return ctx.handled();
-                    }
-
-                    if (ctx.getCreature() != null && ctx.getCreature().isInBattle()) {
-                        ctx.receive(SeeEvent.getBuilder()
-                                .setDeniedReason("You are in a fight right now, you are too busy to examine that!")
-                                .Build());
-                        return ctx.handled();
-                    }
-
-                    for (Item ro : items) {
-                        if (ro.CheckNameRegex(name, 3)) {
-                            ctx.receive(ro.produceMessage(SeeEvent.getBuilder().setExaminable(ro)
-                                    .addExtraInfo("You see it in the room with you. ")));
-                            return ctx.handled();
-                        }
-                    }
-
-                    if (ctx.getCreature() != null) {
-                        ICreature creature = ctx.getCreature();
-                        for (Item thing : creature.getEquipmentSlots().values()) {
-                            if (thing.CheckNameRegex(name, 3)) {
-                                if (thing instanceof Examinable) {
-                                    ctx.receive(((SeeEvent.Builder) thing.produceMessage().copyBuilder())
-                                            .addExtraInfo("You have it equipped. ").Build());
-                                    return ctx.handled();
-                                }
-                                ctx.receive(SeeEvent.getBuilder().setExaminable(thing)
-                                        .addExtraInfo("You have it equipped. ").Build());
-                                return ctx.handled();
-                            }
-                        }
-
-                        Optional<Item> maybeThing = creature.getInventory().getItem(name);
-                        if (maybeThing.isPresent()) {
-                            Item thing = maybeThing.get();
-                            if (thing instanceof Examinable) {
-                                ctx.receive(((SeeEvent.Builder) thing.produceMessage().copyBuilder())
-                                        .addExtraInfo("You see it in your inventory. ").Build());
-                                return ctx.handled();
-                            }
-                            ctx.receive(SeeEvent.getBuilder().setExaminable(thing)
-                                    .addExtraInfo("You see it in your inventory. ").Build());
-                            return ctx.handled();
-                        }
-                    }
-
-                    ctx.receive(
-                            SeeEvent.getBuilder().setDeniedReason("You couldn't find " + name + " to examine. ")
-                                    .Build());
-                    return ctx.handled();
-                } else {
-                    ctx.receive(Room.this.produceMessage());
-                    return ctx.handled();
-                }
-            }
-            return ctx.failhandle();
-        }
-
-        @Override
-        public CommandChainHandler getChainHandler() {
-            return Room.this;
-        }
-    }
-
-    protected class SayHandler implements RoomCommandHandler {
-        private static final String helpString = new StringJoiner(" ")
-                .add("\"say [message]\"").add("Tells everyone in your current room your message").add("\r\n")
-                .add("\"say [message] to [name]\"")
-                .add("Will tell a specific person somewhere in your current room your message.")
-                .add("If your message contains the word 'to', put your message in quotes like")
-                .add("\"say 'They are taking the hobbits to Isengard' to Aragorn\"")
-                .add("\r\n")
-                .toString();
-
-        @Override
-        public AMessageType getHandleType() {
-            return AMessageType.SAY;
-        }
-
-        @Override
-        public Optional<String> getHelp(CommandContext ctx) {
-            return Optional.of(SayHandler.helpString);
-        }
-
-        @Override
-        public Predicate<CommandContext> getEnabledPredicate() {
-            return SayHandler.defaultRoomPredicate;
-        }
-
-        @Override
-        public Reply handleCommand(CommandContext ctx, Command cmd) {
-            if (cmd != null && cmd.getType() == AMessageType.SAY && cmd instanceof SayMessage sMessage) {
-                SpeakingEvent.Builder speakMessage = SpeakingEvent.getBuilder().setSayer(ctx.getCreature())
-                        .setMessage(sMessage.getMessage());
-                if (sMessage.getTarget() != null) {
-                    boolean sent = false;
-                    Optional<ICreature> optTarget = Room.this.getCreature(sMessage.getTarget());
-                    if (optTarget.isPresent()) {
-                        CommandInvoker sayer = ctx.getClient();
-                        if (ctx.getCreature() != null) {
-                            sayer = ctx.getCreature();
-                        } else if (ctx.getUser() != null) {
-                            sayer = ctx.getUser();
-                        }
-                        ICreature target = optTarget.get();
-                        speakMessage.setSayer(sayer).setHearer(target);
-                        ICreature.eventAccepter.accept(target, speakMessage.Build());
-                        sent = true;
-                    }
-                    if (!sent) {
-                        ctx.receive(BadSpeakingTargetEvent.getBuilder().setCreatureName(sMessage.getTarget()));
-                    }
-                } else {
-                    Room.this.announce(speakMessage.Build());
-                }
-                return ctx.handled();
-            }
-            return ctx.failhandle();
-        }
-
-        @Override
-        public CommandChainHandler getChainHandler() {
-            return Room.this;
-        }
-    }
-
-    protected class UseHandler implements RoomCommandHandler {
-        private final static String helpString = new StringJoiner(" ")
-                .add("\"use [itemname]\"").add("Uses an item that you have on yourself, if applicable.")
-                .add("Like \"use potion\"").add("\r\n")
-                .add("\"use [itemname] on [otherthing]\"")
-                .add("Uses an item that you have on something or someone else, if applicable.")
-                .add("Like \"use potion on Bob\"")
-                .toString();
-        private final static Predicate<CommandContext> enabledPredicate = UseHandler.defaultRoomPredicate.and(
-                ctx -> ctx.getCreature().getItems().stream().anyMatch(item -> item != null && item instanceof Usable));
-
-        @Override
-        public AMessageType getHandleType() {
-            return AMessageType.USE;
-        }
-
-        @Override
-        public Optional<String> getHelp(CommandContext ctx) {
-            return Optional.of(UseHandler.helpString);
-        }
-
-        @Override
-        public Predicate<CommandContext> getEnabledPredicate() {
-            return UseHandler.enabledPredicate;
-        }
-
-        @Override
-        public Reply handleCommand(CommandContext ctx, Command cmd) {
-            if (cmd != null && cmd.getType() == AMessageType.USE && cmd instanceof UseMessage useMessage) {
-                if (ctx.getCreature() == null) {
-                    ctx.receive(BadMessageEvent.getBuilder().setBadMessageType(BadMessageType.CREATURES_ONLY)
-                            .setHelps(ctx.getHelps()).setCommand(useMessage).Build());
-                    return ctx.handled();
-                }
-                Optional<Item> maybeItem = ctx.getCreature().getItem(useMessage.getUsefulItem());
-                if (maybeItem.isEmpty() || !(maybeItem.get() instanceof Usable)) {
-                    ctx.receive(ItemUsedEvent.getBuilder().setSubType(UseOutMessageOption.NO_USES)
-                            .setItemUser(ctx.getCreature()).Build());
-                    return ctx.handled();
-                }
-                Usable usable = (Usable) maybeItem.get();
-                if (useMessage.getTarget() == null || useMessage.getTarget().isBlank()) {
-                    usable.doUseAction(ctx, ctx.getCreature());
-                    return ctx.handled();
-                }
-                Collection<ICreature> maybeCreature = Room.this.getCreaturesLike(useMessage.getTarget());
-                if (maybeCreature.size() == 1) {
-                    List<ICreature> creatureList = new ArrayList<>(maybeCreature);
-                    ICreature targetCreature = creatureList.get(0);
-                    // if we aren't in battle, but our target is in battle, join the battle
-                    if (!ctx.getCreature().isInBattle() && targetCreature.isInBattle()) {
-                        final SubArea subArea = Room.this.getSubAreaForSort(SubAreaSort.BATTLE);
-                        if (subArea == null) {
-                            this.log(Level.SEVERE, String.format(
-                                    "How can we target someone in battle without the Room having a battle sub area? %s",
-                                    Room.this.getSubAreas()));
-                            return ctx.failhandle();
-                        }
-                        subArea.addCreature(ctx.getCreature());
-                        return subArea.handleChain(ctx, cmd);
-                    }
-                    usable.doUseAction(ctx, creatureList.get(0));
-                    return ctx.handled();
-                } else if (maybeCreature.size() > 1) {
-                    ctx.receive(BadTargetSelectedEvent.getBuilder().setBde(BadTargetOption.UNCLEAR)
-                            .setBadTarget(useMessage.getTarget()).setPossibleTargets(maybeCreature).Build());
-                    return ctx.handled();
-                }
-                Optional<Item> maybeRoomItem = Room.this.getItem(useMessage.getTarget());
-                if (maybeRoomItem.isPresent()) {
-                    usable.doUseAction(ctx, maybeRoomItem.get());
-                    return ctx.handled();
-                }
-                Optional<Item> maybeInventory = ctx.getCreature().getItem(useMessage.getTarget());
-                if (maybeInventory.isPresent()) {
-                    usable.doUseAction(ctx, maybeInventory.get());
-                    return ctx.handled();
-                }
-                ctx.receive(BadTargetSelectedEvent.getBuilder().setBde(BadTargetOption.UNCLEAR)
-                        .setBadTarget(useMessage.getTarget()).Build());
-                return ctx.handled();
-            }
-            return ctx.failhandle();
-        }
-
-        @Override
-        public CommandChainHandler getChainHandler() {
-            return Room.this;
-        }
-    }
-
-    protected class RestHandler implements RoomCommandHandler {
-        private final static String helpString = "\"REST\" puts yourself in state of REST, use \"GO UP\" to get out of it";
-        private final static Predicate<CommandContext> enabledPredicate = RestHandler.defaultRoomPredicate
-                .and(ctx -> ctx.getArea().hasSubAreaSort(SubAreaSort.RECUPERATION));
-
-        @Override
-        public AMessageType getHandleType() {
-            return AMessageType.REST;
-        }
-
-        @Override
-        public Optional<String> getHelp(CommandContext ctx) {
-            return Optional.of(RestHandler.helpString);
-        }
-
-        @Override
-        public Predicate<CommandContext> getEnabledPredicate() {
-            return RestHandler.enabledPredicate;
-        }
-
-        @Override
-        public Reply handleCommand(CommandContext ctx, Command cmd) {
-            if (cmd == null || cmd.getType() != AMessageType.REST) {
-                return ctx.failhandle();
-            }
-            ctx = Room.this.addSelfToContext(ctx);
-            if (ctx.getCreature() == null) {
-                ctx.receive(BadMessageEvent.getBuilder().setBadMessageType(BadMessageType.CREATURES_ONLY)
-                        .setHelps(ctx.getHelps()).setCommand(cmd).Build());
-                return ctx.handled();
-            }
-            final SubArea subArea = Room.this.getSubAreaForSort(SubAreaSort.RECUPERATION);
-            if (subArea == null) {
-                this.log(Level.WARNING, "No rest sub area found!");
-                return ctx.failhandle();
-            }
-            return subArea.handleChain(ctx, cmd);
-        }
-
-        @Override
-        public CommandChainHandler getChainHandler() {
-            return Room.this;
-        }
     }
 
     @Override
