+++ gameserver/src/main/java/com/lhf/messages/in/TakeMessage.java
@@ -1,5 +1,6 @@
 package com.lhf.messages.in;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.StringJoiner;
 
@@ -17,6 +18,10 @@ public class TakeMessage extends CommandAdapter {
             return null;
         }
         return this.getDirects().get(0);
+    }
+
+    public List<String> getTargets() {
+        return this.getDirects();
     }
 
     public Optional<String> fromContainer() {
